retirando a implementação do login e register do service e colocando no controller. Deixando o service com o UserDetailsService
