+++ src/main/java/com/simplicity/wallet/digital/SimplicityDigitalWallet/config/SecurityConfiguration.java
@@ -1,5 +1,8 @@
 package com.simplicity.wallet.digital.SimplicityDigitalWallet.config;
 
+import com.simplicity.wallet.digital.SimplicityDigitalWallet.service.impl.AuthorizationServiceImpl;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -19,12 +22,14 @@ import org.springframework.security.web.authentication.UsernamePasswordAuthentic
 @EnableWebSecurity
 public class SecurityConfiguration {
 
+    public static Logger logger = LoggerFactory.getLogger(SecurityConfiguration.class);
+
     @Autowired
     private SecurityFilter securityFilter;
 
     @Bean
     public SecurityFilterChain configure(HttpSecurity http) throws Exception {
-
+        logger.info("Configuring HttpSecurity");
         return http
                 .csrf(AbstractHttpConfigurer::disable)
                 .sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
@@ -40,11 +45,14 @@ public class SecurityConfiguration {
 
     @Bean
     public AuthenticationManager authenticationManager(AuthenticationConfiguration authenticationConfiguration) throws Exception {
+        logger.info("Configuring AuthenticationManager");
         return authenticationConfiguration.getAuthenticationManager();
     }
 
     @Bean
     public PasswordEncoder passwordEncoder() {
+        logger.info("Configuring PasswordEncoder");
         return new BCryptPasswordEncoder();
     }
+
 }
