+++ src/main/java/com/simplicity/wallet/digital/SimplicityDigitalWallet/config/SecurityFilter.java
@@ -48,6 +48,7 @@ public class SecurityFilter  extends OncePerRequestFilter {
     }
 
     private String recoverToken(HttpServletRequest request){
+        logger.info("Inicio recoverToken");
         var authHeader = request.getHeader("Authorization");
         if(authHeader == null) return null;
         return authHeader.replace("Bearer ", "");
